Improve: refactored so that Gallery only depends on content, 'category' is now derived, and it no longer knows about `playMode`
