+++ src/components/Gallery/styled.ts
@@ -42,7 +42,7 @@ export const Grid = styled.div`
   `}
 `;
 
-export const Container = styled.div`
+export const Section = styled.section`
   display: flex;
   justify-content: center;
   align-items: center;
